+++ src/day9.rs
@@ -1,9 +1,9 @@
 use crate::computer::{parse_memory, Computer};
 
-pub fn first(contents: &String) -> i128 {
+pub fn first(contents: &String) -> i64 {
     Computer::run(parse_memory(contents), vec![1]).output[0]
 }
 
-pub fn second(contents: &String) -> i128 {
+pub fn second(contents: &String) -> i64 {
     Computer::run(parse_memory(contents), vec![2]).output[0]
 }
